+++ frontend/src/app/modules/page/notifications/detail/edit/notification-edit.component.ts
@@ -199,6 +199,14 @@ export class NotificationEditComponent implements OnDestroy {
 
   }
 
+  paginationChangedAffectedParts(event: any){
+    this.setAffectedPartsBasedOnNotificationType(this.selectedNotification, this.cachedAffectedPartsFilter, event);
+  }
+
+  paginationChangedAvailableParts(event: any){
+    this.setAvailablePartsBasedOnNotificationType(this.selectedNotification, this.cachedAvailablePartsFilter, event);
+  }
+
   filterAvailableParts(partsFilter: any): void {
     if (this.cachedAvailablePartsFilter !== JSON.stringify(partsFilter)) {
       this.setAvailablePartsBasedOnNotificationType(this.selectedNotification, partsFilter);
@@ -229,25 +237,25 @@ export class NotificationEditComponent implements OnDestroy {
   }
 
 
-  private setAvailablePartsBasedOnNotificationType(notification: Notification, assetFilter?: any) {
+  private setAvailablePartsBasedOnNotificationType(notification: Notification, assetFilter?: any, pagination?: any) {
     if (this.affectedPartIds) {
       assetFilter = this.enrichPartsFilterByAffectedAssetIds(null, true);
     }
     if (notification.type === NotificationType.INVESTIGATION) {
-      this.partsFacade.setSupplierPartsAsBuilt(FIRST_PAGE, DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(assetFilter, true));
+      this.partsFacade.setSupplierPartsAsBuilt(pagination?.page || FIRST_PAGE, pagination?.pageSize || DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(assetFilter, true));
     } else {
-      this.ownPartsFacade.setPartsAsBuilt(FIRST_PAGE, DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(assetFilter, true));
+      this.ownPartsFacade.setPartsAsBuilt(pagination?.page || FIRST_PAGE, pagination?.pageSize || DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(assetFilter, true));
     }
   }
 
-  private setAffectedPartsBasedOnNotificationType(notification: Notification, partsFilter?: any) {
+  private setAffectedPartsBasedOnNotificationType(notification: Notification, partsFilter?: any, pagination?: any) {
 
     if (this.affectedPartIds.length > 0) {
       partsFilter = this.enrichPartsFilterByAffectedAssetIds(null);
       if (notification.type === NotificationType.INVESTIGATION) {
-        this.partsFacade.setSupplierPartsAsBuiltSecond(FIRST_PAGE, DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(partsFilter, true));
+        this.partsFacade.setSupplierPartsAsBuiltSecond(pagination?.page || FIRST_PAGE, pagination?.pageSize || DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(partsFilter, true));
       } else {
-        this.ownPartsFacade.setPartsAsBuiltSecond(FIRST_PAGE, DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(partsFilter, true));
+        this.ownPartsFacade.setPartsAsBuiltSecond(pagination?.page || FIRST_PAGE, pagination?.pageSize || DEFAULT_PAGE_SIZE, this.tableAsBuiltSortList, toAssetFilter(partsFilter, true));
       }
     } else {
       this.partsFacade.setSupplierPartsAsBuiltSecondEmpty();
@@ -298,7 +306,9 @@ export class NotificationEditComponent implements OnDestroy {
     this.isSaveButtonDisabled = this.notificationFormGroup.invalid || this.affectedPartIds.length < 1;
     this.deselectPartTrigger$.next(this.temporaryAffectedParts);
     this.currentSelectedAvailableParts$.next([]);
+/*    this.currentSelectedAffectedParts$.next([]);*/
     this.temporaryAffectedParts = [];
+    console.log(this.selectedNotification, "selected");
     this.setAffectedPartsBasedOnNotificationType(this.selectedNotification);
     this.setAvailablePartsBasedOnNotificationType(this.selectedNotification);
   }
@@ -321,6 +331,9 @@ export class NotificationEditComponent implements OnDestroy {
       .getNotificationById(notificationId)
       .subscribe({
         next: data => {
+          console.log(data.type, "data.type from next");
+          console.log(data, "data from next");
+
           this.selectNotificationAndLoadPartsBasedOnNotification(data);
         },
         error: () => {
@@ -330,7 +343,9 @@ export class NotificationEditComponent implements OnDestroy {
   }
 
   private selectNotificationAndLoadPartsBasedOnNotification(notification: Notification) {
+    console.log(notification, "selected from selectNotificationAndLoadPartsBasedOnNotification");
     this.selectedNotification = notification;
+
     this.affectedPartIds = notification.assetIds;
     this.tableType = notification.type === NotificationType.INVESTIGATION ? TableType.AS_BUILT_SUPPLIER : TableType.AS_BUILT_OWN;
     this.setAvailablePartsBasedOnNotificationType(this.selectedNotification);